fastapi_server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0,<2.12
loguru>=0.7.0
python-dotenv>=1.0.0
langchain==1.0
langchain-core>=0.1.0
langchain-community
langchain-openai
langchain-mcp-adapters
langgraph>=0.0.40
tyro>=0.7.0
commentjson>=0.9.0
matplotlib>=3.7.0
Pillow>=10.0.0
jax>=0.4.0
httpx[socks]
dashscope
websockets>=11.0.3
mcp>=1.8.1
mcp-proxy>=0.8.2
faiss-cpu
fastmcp
pandas
